+++ javers-persistence-mongo/src/main/java/org/javers/repository/mongo/MongoRepository.java
@@ -27,46 +27,28 @@ import org.javers.repository.api.QueryParamsBuilder;
 import org.javers.repository.api.SnapshotIdentifier;
 import org.javers.repository.mongo.model.MongoHeadId;
 import org.joda.time.LocalDateTime;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.*;
 
 import static org.javers.common.validation.Validate.conditionFulfilled;
+import static org.javers.repository.mongo.MongoSchemaManager.*;
 
 /**
  * @author pawel szymczyk
  */
 public class MongoRepository implements JaversRepository {
-    private static final Logger logger = LoggerFactory.getLogger(MongoRepository.class);
 
     private static final int DESC = -1;
-    private static final int ASC = 1;
-
-    private static final String SNAPSHOTS = "jv_snapshots";
-    private static final String COMMIT_ID = "commitMetadata.id";
-    private static final String COMMIT_DATE = "commitMetadata.commitDate";
-    private static final String COMMIT_AUTHOR = "commitMetadata.author";
-    private static final String COMMIT_PROPERTIES = "commitMetadata.properties";
-    private static final String GLOBAL_ID_KEY = "globalId_key";
-    private static final String GLOBAL_ID_ENTITY = "globalId.entity";
-    private static final String GLOBAL_ID_OWNER_ID_ENTITY = "globalId.ownerId.entity";
-    private static final String GLOBAL_ID_FRAGMENT = "globalId.fragment";
-    private static final String GLOBAL_ID_VALUE_OBJECT = "globalId.valueObject";
-    private static final String SNAPSHOT_VERSION = "version";
-    private static final String CHANGED_PROPERTIES = "changedProperties";
-    private static final String OBJECT_ID = "_id";
-
-    private MongoDatabase mongo;
+    private final MongoSchemaManager mongoSchemaManager;
     private JsonConverter jsonConverter;
 
     public MongoRepository(MongoDatabase mongo) {
-        this.mongo = mongo;
+        this(mongo, null);
     }
 
     MongoRepository(MongoDatabase mongo, JsonConverter jsonConverter) {
-        this.mongo = mongo;
         this.jsonConverter = jsonConverter;
+        this.mongoSchemaManager = new MongoSchemaManager(mongo);
     }
 
     @Override
@@ -139,41 +121,7 @@ public class MongoRepository implements JaversRepository {
 
     @Override
     public void ensureSchema() {
-        //ensures collections and indexes
-        MongoCollection<Document> snapshots = snapshotsCollection();
-        snapshots.createIndex(new BasicDBObject(GLOBAL_ID_KEY, ASC));
-        snapshots.createIndex(new BasicDBObject(GLOBAL_ID_VALUE_OBJECT, ASC));
-        snapshots.createIndex(new BasicDBObject(GLOBAL_ID_OWNER_ID_ENTITY, ASC));
-        snapshots.createIndex(new BasicDBObject(CHANGED_PROPERTIES, ASC));
-        snapshots.createIndex(new BasicDBObject(COMMIT_PROPERTIES + ".key", ASC).append(COMMIT_PROPERTIES + ".value", ASC));
-
-        headCollection();
-
-        //schema migration script from JaVers 2.0 to 2.1
-        dropIndexIfExists(snapshots, GLOBAL_ID_ENTITY);
-
-        //schema migration script from JaVers 1.1 to 1.2
-        Document doc = snapshots.find().first();
-        if (doc != null) {
-            Object stringCommitId = ((Map)doc.get("commitMetadata")).get("id");
-            if (stringCommitId instanceof String) {
-                logger.info("executing db migration script, from JaVers 1.1 to 1.2 ...");
-
-                Document update = new Document("eval",
-                    "function() { \n"+
-                            "    db.jv_snapshots.find().forEach( \n"+
-                            "      function(snapshot) { \n"+
-                            "        snapshot.commitMetadata.id = Number(snapshot.commitMetadata.id); \n"+
-                            "        db.jv_snapshots.save(snapshot); } \n" +
-                            "    ); "+
-                            "    return 'ok'; \n"+
-                            "}"
-                    );
-
-                Document ret = mongo.runCommand(update);
-                logger.info("result: \n "+ ret.toJson());
-            }
-        }
+        mongoSchemaManager.ensureSchema();
     }
 
     private Bson createIdQuery(GlobalId id) {
@@ -227,11 +175,11 @@ public class MongoRepository implements JaversRepository {
     }
 
     private MongoCollection<Document> snapshotsCollection() {
-        return mongo.getCollection(SNAPSHOTS);
+        return mongoSchemaManager.snapshotsCollection();
     }
 
     private MongoCollection<Document> headCollection() {
-        return mongo.getCollection(MongoHeadId.COLLECTION_NAME);
+        return mongoSchemaManager.headCollection();
     }
 
     private void persistSnapshots(Commit commit) {
@@ -377,19 +325,5 @@ public class MongoRepository implements JaversRepository {
     //enables index range scan
     private static Bson prefixQuery(String fieldName, String prefix){
         return Filters.regex(fieldName, "^" + RegexEscape.escape(prefix) + ".*");
-    }
-
-    private static void dropIndexIfExists(MongoCollection<Document> col, String fieldName){
-        for (Document d : col.listIndexes()) {
-            if (d.get("key", Document.class).containsKey(fieldName)) {
-                logger.warn("Schema migration. Dropping index: "+fieldName +" ...");
-
-                try {
-                    col.dropIndex(fieldName);
-                }catch (Exception swallowed){} //because not implemented in Fongo
-
-                return;
-            }
-        }
     }
 }
